Handle local time being different from twitter's
The twitter hourly limit includes a "reset" time (unix timestamp) for the
current count, which is occasionally *before* the local timestamp. This
previously resulted in negative time deltas and Twitux stiopped updating.
We now catch this case and assume the seconds are 1, as the reset time is
generally very close in this case.