Some images are semi-invalid, so don't break on them
Twitter right now returns all profile images claiming to be JPGs, when some of
them are in fact PNGs, which confuses the crap out of
gdk_pixbuf_new_from_file (certainly in Gtk 2.14) which seems to do most of
its sniffing based on filenames... These then fail to load, and Twitux
complains lots. This patch catches a couple more of those cases.